Improve colors file searching on Windows
On Windows, minetestmapper now uses the colors files from the
directory where it is installed.

If the last directory is named 'bin', that part is stripped,
and the resulting directory, and a subdirectory named 'colors'
are searched.

All depends on the whether on invocation, argv[0] contains the full
path of the exe file. It works under Wine when compiled with MinGW.

I haven't been able to test on a real copy of windows, or when
compiled using msvc, but I assume that will work too.